Improved UI interface for pending and applied migrations & moved runner actions to a separated controller
