+++ templates/log.twig
@@ -1,32 +1,18 @@
-{% extends "_layouts/cp" %}
-{% set title = "Migration Manager"|t %}
+{% extends 'migrationManager/_layouts/cp' %}
+
+{% set title = 'Migration Logs'|t %}
+
+{% set crumbs = [
+    { label: craft.migrationManager.name, url: url('migrationmanager') },
+    { label: 'Migration Logs'|t, url: url('migrationmanager/logs') }
+] %}
 
 {% import "_includes/forms" as forms %}
 
-{% includeCssResource "migrationmanager/css/migration-manager.css" %}
-
-
-{% set tabs = {
-    tab1: { label: "Create Migration"|t, url: url('migrationmanager') },
-    tab2: { label: "Migrations"|t, url: url('migrationmanager/migrations') },
-    tab3: {label: "Log"|t, url: url('migrationmanager/log')}
-} %}
-
-{% set selectedTab = 'tab3' %}
-
 {% set content %}
-    <div class="migration-manager">
-        <h1>Migration Log</h1>
-
-        {% for logEntry in logEntries %}
-            <h3>{{ logEntry.dateTime.format('Y-m-d H:i:s') }}</h3>
-            <p><span class="error-type">{{ logEntry.level }}</span>&nbsp;{{ logEntry.message }}</p>
-            <hr/>
-        {% endfor %}
-
-
-
-
-    </div>
-
+    {% for logEntry in logEntries %}
+        <h3>{{ logEntry.dateTime.format('Y-m-d H:i:s') }}</h3>
+        <p><span class="error-type">{{ logEntry.level|upper }}</span>&nbsp;{{ logEntry.message }}</p>
+        {% if not loop.last %}<hr>{% endif %}
+    {% endfor %}
 {% endset %}
