--- templates/migrations-applied.twig
@@ -1,50 +1,0 @@
-{% extends "_layouts/cp" %}
-{% set title = "Migration Manager"|t %}
-
-{% includeCssResource "migrationmanager/css/styles.css" %}
-
-{% import "_includes/forms" as forms %}
-
-{% set tabs = {
-    tab1: { label: "Create Migration"|t, url: url('migrationmanager') },
-    tab2: { label: "Migrations"|t, url: url('migrationmanager/migrations') },
-    tab3: {label: "Log"|t, url: url('migrationmanager/log')}
-} %}
-
-{% set selectedTab = 'tab2' %}
-
-{% set content %}
-<div class="migration-manager">
-    {% set migrations = craft.migrationManager.getAppliedMigrations() %}
-    {%  if migrations | length %}
-        <h1>Applied migration{{ migrations|length > 1 ? 's' : '' }}</h1>
-    {%  else %}
-        <h1>No applied migrations</h1>
-    {%  endif %}
-
-    <form class="field toolbar" action="" method="post" accept-charset="UTF-8">
-        <input type="hidden" name="action" value="migrationManager/runAppliedMigration">
-        <input type="hidden" name="applied" value="1">
-        {{ getCsrfInput() }}
-
-        {% for migration in migrations %}
-            <input type="checkbox" id="{{ migration }}" class="checkbox" name="migration[]" value="{{ migration }}"> <label for="{{ migration }}">{{ migration }}</label>
-            <hr/>
-        {% endfor %}
-
-        {% if migrations|length %}
-            <p>You must select specific migrations to run. Return to the <a class="link-muted" href="migrations">new migrations</a>.</p>
-            <input type="submit" class="btn submit export-btn" value="Rerun Migration{{ migrations|length > 1 ? 's' : '' }}">
-        {% else %}
-            <p>There are no applied migrations. Return to the <a class="link-muted" href="migrations">new migrations</a>.</p>
-        {% endif %}
-
-    </form>
-
-
-
-
-
-</div>
-
-{% endset %}
