+++ templates/pending.twig
@@ -1,0 +1,43 @@
+{% extends 'migrationManager/_layouts/cp' %}
+
+{% set title = 'Pending Migrations'|t %}
+
+{% set crumbs = [
+    { label: craft.migrationManager.name, url: url('migrationmanager') },
+    { label: 'Pending Migrations'|t, url: url('migrationmanager/pending') }
+] %}
+
+{% block content %}
+
+    {% set migrations = craft.migrationManager.getNewMigrations() %}
+    {% if migrations | length > 0%}
+        <p>
+            {{ 'Select the migration you want to run.'|t }}
+        </p>
+
+        <form class="field toolbar" action="" method="post" accept-charset="UTF-8">
+            <input type="hidden" name="action" value="migrationManager/run/start">
+            {{ getCsrfInput() }}
+
+            {% for migration in migrations %}
+                <input type="checkbox" id="{{ migration }}" class="checkbox" name="migration[]" value="{{ migration }}">
+                <label for="{{ migration }}">{{ migration }}</label>
+                {% if not loop.last %}<hr>{% endif %}
+            {% endfor %}
+
+            <div class="buttons clearafter">
+                <button class="btn submit export-btn" name="submit" value="create">
+                    {{ 'Run migrations'|t }}
+                </button>
+            </div>
+
+            <div class="warning">
+                {{ 'If you don\'t make a selection, all migrations will be run.'|t }}
+            </div>
+        </form>
+    {% else %}
+        <p>
+            {{ 'There are no pending migrations. You can run previously <a href="applied">applied migrations</a>.'|t|raw }}
+        </p>
+    {% endif %}
+{% endblock %}
